Implement input array preprocessing in buildTree method of Tree class
First, duplicates are removed from the array, as requested.
The reason is that they make the building of the balanced binary tree
more complicated and result in trees that are much harder to balance.

Second, the array is sorted. Sorting an array can be, eg, O(n logn)
if using merge sort. The reason is that then you can build a balanced
binary tree in O(n) starting from a sorted array.